+++ src/compiler/component-hydrate/write-hydrate-outputs.ts
@@ -10,7 +10,6 @@ export function writeHydrateOutputs(config: d.Config, compilerCtx: d.CompilerCtx
 
 
 async function writeHydrateOutput(config: d.Config, compilerCtx: d.CompilerCtx, buildCtx: d.BuildCtx, outputTarget: d.OutputTargetHydrate, rollupOutput: RollupOutput) {
-  const hydrateAppFileName = getHydrateAppFileName(config);
   const hydratePackageName = await getHydratePackageName(config, compilerCtx);
 
   const hydrateAppDirPath = outputTarget.dir;
@@ -21,22 +20,28 @@ async function writeHydrateOutput(config: d.Config, compilerCtx: d.CompilerCtx, 
   const pkgJsonPath = config.sys.path.join(hydrateAppDirPath, 'package.json');
   const pkgJsonCode = getHydratePackageJson(config, hydrateCoreIndexPath, hydrateCoreIndexDtsFilePath, hydratePackageName);
 
-  const writePromises: Promise<any>[] = [
-    copyHydrateRunner(config, compilerCtx, hydrateAppDirPath, hydrateAppFileName, hydratePackageName),
+  await Promise.all([
+    copyHydrateRunnerDts(config, compilerCtx, hydrateAppDirPath),
     compilerCtx.fs.writeFile(pkgJsonPath, pkgJsonCode)
-  ];
-
-  rollupOutput.output.forEach(output => {
-    if (output.type === 'chunk') {
-      const filePath = config.sys.path.join(hydrateAppDirPath, output.fileName);
-      writePromises.push(compilerCtx.fs.writeFile(filePath, output.code));
-    }
-  });
+  ]);
 
   // always remember a path to the hydrate app that the prerendering may need later on
   buildCtx.hydrateAppFilePath = hydrateCoreIndexPath;
 
-  return Promise.all(writePromises);
+  await Promise.all(rollupOutput.output.map(async output => {
+    if (output.type === 'chunk') {
+      const filePath = config.sys.path.join(hydrateAppDirPath, output.fileName);
+      try {
+        const existingCode = await compilerCtx.fs.disk.readFile(filePath);
+        if (existingCode === output.code) {
+          // if it's identical then don't overwrite it so debugging context works
+          return;
+        }
+      } catch (e) {}
+
+      await compilerCtx.fs.writeFile(filePath, output.code);
+    }
+  }));
 }
 
 
@@ -63,30 +68,14 @@ async function getHydratePackageName(config: d.Config, compilerCtx: d.CompilerCt
 }
 
 
-export function getHydrateAppFileName(config: d.Config) {
-  return `${config.fsNamespace}-hydrate.js`;
-}
+async function copyHydrateRunnerDts(config: d.Config, compilerCtx: d.CompilerCtx, hydrateAppDirPath: string) {
+  const srcHydrateDir = config.sys.path.join(config.sys.compiler.distDir, 'hydrate');
 
+  const runnerDtsFileName = 'index.d.ts';
 
-async function copyHydrateRunner(config: d.Config, compilerCtx: d.CompilerCtx, hydrateAppDirPath: string, hydrateAppFileName: string, hydratePackageName: string) {
-  const srcHydrateDir = config.sys.path.join(config.sys.compiler.packageDir, 'internal', 'hydrate');
+  const runnerDtsSrcPath = config.sys.path.join(srcHydrateDir, runnerDtsFileName);
 
-  const runnerSrcPath = config.sys.path.join(srcHydrateDir, 'runner.js');
-  const runnerDtsSrcPath = config.sys.path.join(srcHydrateDir, 'runner.d.ts');
+  const runnerDtsDestPath = config.sys.path.join(hydrateAppDirPath, runnerDtsFileName);
 
-  const appRunnerIndexFileName = 'index.js';
-  const appRunnerDtsFileName = 'index.d.ts';
-
-  const runnerDestPath = config.sys.path.join(hydrateAppDirPath, appRunnerIndexFileName);
-  const runnerDtsDestPath = config.sys.path.join(hydrateAppDirPath, appRunnerDtsFileName);
-
-  let runnerSrcCode = await compilerCtx.fs.readFile(runnerSrcPath);
-
-  runnerSrcCode = runnerSrcCode.replace('$$HYDRATE_APP_FILENAME$$', hydrateAppFileName);
-  runnerSrcCode = runnerSrcCode.replace('$$HYDRATE_APP_PACKAGE_NAME$$', hydratePackageName);
-
-  await Promise.all([
-    compilerCtx.fs.writeFile(runnerDestPath, runnerSrcCode),
-    compilerCtx.fs.copyFile(runnerDtsSrcPath, runnerDtsDestPath)
-  ]);
+  await compilerCtx.fs.copyFile(runnerDtsSrcPath, runnerDtsDestPath);
 }