+++ src/hydrate/platform/index.ts
@@ -129,6 +129,6 @@ export const Build: d.UserBuildConditionals = {
 export const styles: d.StyleMap = new Map();
 
 export { BUILD, NAMESPACE, globalScripts } from '@app-data';
-export { bootstrapHydrate } from './bootstrap-hydrate';
+export { hydrateApp } from './hydrate-app';
 
 export * from '@runtime';