+++ src/runtime/event-emitter.ts
@@ -10,16 +10,16 @@ export const createEvent = (ref: d.RuntimeRef, name: string, flags: number) => {
   return {
     emit: (detail: any) => {
       if (BUILD.isDev && !elm.isConnected) {
-        consoleDevWarn(`The "${name}" event was emitted, but the dispatcher node is not longer connected to the dom.`);
+        consoleDevWarn(`The "${name}" event was emitted, but the dispatcher node is no longer connected to the dom.`);
       }
-      return elm.dispatchEvent(
-        new (BUILD.hydrateServerSide ? (win as any).CustomEvent : CustomEvent)(name, {
-          bubbles: !!(flags & EVENT_FLAGS.Bubbles),
-          composed: !!(flags & EVENT_FLAGS.Composed),
-          cancelable: !!(flags & EVENT_FLAGS.Cancellable),
-          detail
-        })
-      );
+      const ev = new (BUILD.hydrateServerSide ? (win as any).CustomEvent : CustomEvent)(name, {
+        bubbles: !!(flags & EVENT_FLAGS.Bubbles),
+        composed: !!(flags & EVENT_FLAGS.Composed),
+        cancelable: !!(flags & EVENT_FLAGS.Cancellable),
+        detail
+      });
+      elm.dispatchEvent(ev);
+      return ev;
     }
   };
 };