+++ test/hello-world/stencil.config.ts
@@ -3,6 +3,8 @@ import { Config } from '../../internal';
 export const config: Config = {
   namespace: 'HelloWorld',
   outputTargets: [
+    { type: 'dist' },
+    { type: 'dist-hydrate-script' },
     {
       type: 'www',
       serviceWorker: null,